split salesforce task file into class files
