+++ cumulusci/tasks/salesforce/BaseSalesforceMetadataApiTask.py
@@ -1,0 +1,14 @@
+from cumulusci.tasks.salesforce import BaseSalesforceTask
+
+
+class BaseSalesforceMetadataApiTask(BaseSalesforceTask):
+    api_class = None
+    name = 'BaseSalesforceMetadataApiTask'
+
+    def _get_api(self):
+        return self.api_class(self)
+
+    def _run_task(self):
+        api = self._get_api()
+        if api:
+            return api()
